+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## What's New in 3.0.10
+
+### Bugs fixed
+
+-   Fixed bug that caused comments in the PR diff view to stop showing up.
+
+### Improvements
+
+-   Fixed Incomplete List of Disallowed Inputs vulnerability in affected versions of babel/traverse.
+-   Fixed Prototype Pollution vulnerability in affected versions of axios
+
 ## What's New in 3.0.9
 
 ### Bugs fixed